+++ CHANGELOG.md
@@ -19,6 +19,12 @@
 * Switched our async runtimes integration from `executor-trait` and `reactor-trait` to `async-rs`
 * Update amq-protocol to 10.0
 
+### 3.7.1 (2025-10-12)
+
+#### Bug Fixes
+
+* Fix `io_loop` tracing span
+
 ### 3.7.0 (2025-09-22)
 
 #### Bug Fixes
@@ -126,6 +132,12 @@
 * Drop parking-lot dependency
 * Drop pinky-swear dependency
 * Edition 2024
+
+### 2.5.5 (2025-10-12)
+
+#### Bug Fixes
+
+* Fix `io_loop` tracing span
 
 ### 2.5.4 (2025-07-22)